pyproject.toml
[project]
name = "hr-pro"
version = "0.1.0"
description = "HR AI Agent"
package-mode = false
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-cloud-firestore (>=2.22.0,<3.0.0)",
    "google-cloud-tasks (>=2.20.0,<3.0.0)",
    "tenacity (>=8.2.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "h2 (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
